requirements.txt
Django==5.2.5
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.15.2
django-cors-headers==4.3.1

